feat/todo-app#38: restructure type and constant
