+++ frontend/src/types/task.type.ts
@@ -1,0 +1,3 @@
+export type TTaskStatus = "NOT_STARTED" | "IN_PROGRESS" | "DONE" | "FAILED"
+export type TTaskPriority = "HIGH" | "MEDIUM" | "LOW"
+export type TTaskSortBy = "newest" | "oldest" | "dueDate" | "priority"
